# app/src/main/java/fr/asys/starter/cleher/core/api/taskmanager/dto/TaskDto.java
package fr.asys.starter.cleher.core.api.taskmanager.dto;

import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.Id;

@Entity
public class TaskDto {
    @Id
    @GeneratedValue()
    private int id;
    private String task;
    private boolean done;

    public String getTask() {
        return task;
    }

    public void setTask(String task) {
        this.task = task;
    }

    public boolean isDone() {
        return done;
    }

    public void setDone(boolean done) {
        this.done = done;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }
}